Refactored provenanceFilePicker into it's own component
